cloud/backend/requirements.txt
# IntelliOptics 2.0 - Cloud Backend Dependencies

# Web framework
fastapi>=0.115.0
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Azure integrations (optional — no longer required, Supabase is primary)
# azure-storage-blob==12.19.0
# azure-identity==1.15.0
# azure-servicebus==7.11.4

# Email
sendgrid==6.11.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client
httpx==0.25.0

# Configuration
pydantic>=2.9.2
pydantic-settings>=2.1.0
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0
jinja2>=3.1.0

# SMS
twilio>=9.0.0

# Video processing
yt-dlp==2024.3.10
streamlink>=7.0.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
framegrab[youtube]>=0.14.0
